+++ web/Cargo.toml
@@ -12,7 +12,7 @@ serde_json = "1.0"
 tokio={version="1", features=["full"]}
 log = "0.4"
 pretty_env_logger="0.5"
-postgres = "0.19.7"
+
 
 [profile.release]
 opt-level = 3
