+++ web/src/lib.rs
@@ -1,3 +1,1 @@
 pub mod router;
-pub mod data_struct;
-pub mod database_statement;
