+++ web_api/src/data_struct/mod.rs
@@ -3,8 +3,12 @@ mod sign_in;
 mod user;
 mod sign_in_statement;
 mod sign_up_statements;
+mod upload_profile_photo;
+mod upload_profile_photo_statement;
 pub use sign_up::*;
 pub use sign_in::*;
 pub use user::*;
 pub use sign_in_statement::*;
 pub use sign_up_statements::*;
+pub use upload_profile_photo::*;
+pub use upload_profile_photo_statement::*;
