+++ web_api/src/data_struct/upload_profile_photo_statement.rs
