+++ web_api/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod api_router;
-pub mod upload;
+pub mod data_struct;
+pub mod database_statement;
