--- web_api/src/upload/mod.rs
@@ -1,4 +1,0 @@
-mod upload_profile_photo;
-mod upload_profile_photo_statement;
-pub use upload_profile_photo::*;
-pub use upload_profile_photo_statement::*;
