--- web_api/src/upload/upload_profile_photo.rs
@@ -1,16 +1,0 @@
-
-#[derive(serde::Deserialize,serde::Serialize)]
-pub struct UploadUserProfilePhotoMessage{
-    message:String,
-    message_type:u32,
-    flag:bool,
-}
-impl UploadUserProfilePhotoMessage{
-    pub fn new(message:String,message_type:u32,flag:bool)->Self{
-        UploadUserProfilePhotoMessage{
-            message,
-            message_type,
-            flag,
-        }
-    }
-}
